requirements.txt
absl-py==2.2.2
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
audioread==3.0.1
babel==2.16.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.1.1
cvzone==1.6.1
cycler==0.12.0
Cython==3.0.5
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
executing==2.2.0
facexlib==0.3.0
fastjsonschema==2.21.1
filelock==3.13.1
filterpy==1.4.5
fonttools==4.43.0
fqdn==1.5.1
fsspec==2023.10.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hydra-core==1.3.2
idna==3.4
imageio==2.32.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lapx==0.5.5
lazy_loader==0.3
librosa==0.11.0
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.8
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.7
mistune==3.1.0
mpmath==1.3.0
msgpack==1.1.0
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.0
omegaconf==2.3.0
opencv-python==4.8.1.78
opt_einsum==3.4.0
overrides==7.7.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pgmpy==1.0.0
pickleshare==0.7.5
Pillow==10.0.1
pipreqs==0.5.0
platformdirs==4.3.6
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==6.30.2
psutil==5.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==3.2.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
realesrgan==0.3.0
referencing==0.36.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.12
tb-nightly==2.20.0a20250412
tensorboard-data-server==0.7.2
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.9.26
tinycss2==1.4.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.212
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.29.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
yapf==0.43.0
yarg==0.1.9
